completed stripe payment method

+++ controllers/admin.js
@@ -1,4 +1,6 @@
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
+
 const Product = require('../models/product');
 const { validationResult } = require('express-validator/check');
 
@@ -171,6 +173,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if(image){
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
             return product.save()
@@ -208,17 +211,43 @@ exports.getProducts = (req, res, next) => {
 }
 
 /*====================== Post Delete Product ======================================================================== */
-exports.postDeleteProduct = (req, res, next) => {
-    const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })           
-            .then(() => {
-                console.log('Destroyed Product');
-                res.redirect('/admin/products');
-            })
-            .catch(err => {
-                const error = new Error(err);
-                error.httpStatusCode = 500;
-                return next(error);
-            });
-    
+// exports.postDeleteProduct = (req, res, next) => {
+//     const prodId = req.body.productId;
+//     Product.findById(prodId)
+//     .then(product => {
+//         if(!product){
+//             return next(new Error('Product not found.'))
+//         }
+//         fileHelper.deleteFile(product.imageUrl);
+//         return Product.deleteOne({ _id: prodId, userId: req.user._id })  
+//     })
+//     .then(() => {
+//         console.log('Destroyed Product');
+//         res.redirect('/admin/products');
+//     })
+//     .catch(err => {
+//         const error = new Error(err);
+//         error.httpStatusCode = 500;
+//         return next(error);
+//     }); 
+// }
+
+/*====================== Delete Product ======================================================================== */
+exports.deleteProduct = (req, res, next) => {
+    const prodId = req.params.productId;
+    Product.findById(prodId)
+    .then(product => {
+        if(!product){
+            return next(new Error('Product not found.'))
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({ _id: prodId, userId: req.user._id })  
+    })
+    .then(() => {
+        console.log('Destroyed Product');
+        res.status(200).json({message: 'Success!'});
+    })
+    .catch(err => {
+        res.status(500).json({message: 'Deleting product failed!'});
+    }); 
 }
